Added base files for Spring framework
Spring configuration and a simple "Hello World" page. Configurations
were pulled and trimmed from a previous project of mine. Test on Bluemix
was a success.
Note: the project still needs to be renamed once we have a name for it,
and this can be tricky since the name appears in many locations. It is
currently "SpringMVCTest".