Fixed portal transformations; implemented oblique near clipping plane for portal cameras; scaled down sky box; debug mode improvements
The transform applied to objects interacting with a portal works for all portal angles now, but the code is a bit of a mess. It could still be optimized and improved. Sky box is now at 1/100 its original distance to reduce clipping. I tried a logarithmic z-buffer approach that would totally eliminate clipping, but it messed up a strategy I found for making the near clipping plane perfect for portal cameras. This method was designed for the default z-buffer. If I can manage to adapt it to a logarithmic buffer, we can hope to see our clipping issues go away. Scenes viewed through portals still do not have shadow; I suspect this has something to do with nonzero stencil buffer values. Debug mode now shows the player camera's forward vector and allows you to toggle first and third person with V.